# modules/openapi/src/commonMain/kotlin/com/sphereon/oid/fed/openapi/admin-server.yaml
openapi: 3.0.3
info:
  title: OpenID Federation API
  description: This API allows the interaction with federation endpoints to manage and verify entities within a federation.
  contact:
    name: Sphereon
    email: info@sphereon.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0-d40
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/SphereonInt/OpenIDFederationAPI/1.0.0-d40'
  - url: http://localhost:8081
    description: Admin Server
paths:
...
  '/metadata/{id}':
    delete:
      summary: Delete account metadata
      description: Deletes a specific metadata entry for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the metadata entry to be deleted.
      responses:
        '200':
          description: Metadata deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  '/subordinates':
    get:
      summary: Get subordinates
      description: Retrieve all subordinates associated with the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successfully retrieved subordinates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subordinate'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a subordinate
      description: Create a subordinate for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        description: Subordinate creation details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubordinate'
      responses:
        '201':
          description: Subordinate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subordinate'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{id}':
    delete:
      summary: Delete subordinate
      description: Deletes a specific subordinate for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate to be deleted.
      responses:
        '200':
          description: Subordinate deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Subordinate deleted successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subordinate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{id}/jwks':
    get:
      summary: Get subordinate JWKs
      description: Retrieve all JWKs associated with the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      responses:
        '200':
          description: Successfully retrieved JWKs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubordinateJwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a subordinate JWK
      description: Create a JWK for the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      requestBody:
        description: Metadata creation details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseJwk'
      responses:
        '201':
          description: Subordinate JWK created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateJwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{id}/jwks/{jwkId}':
    delete:
      summary: Delete subordinate JWK
      description: Delete a JWK associated with the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
        - name: jwkId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the JWK.
      responses:
        '200':
          description: Subordinate JWK deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateJwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{subordinateId}/metadata':
    get:
      summary: Get subordinate metadata
      description: Retrieve metadata associated with the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: subordinateId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      responses:
        '200':
          description: Successfully retrieved subordinate metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubordinateMetadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create subordinate metadata
      description: Create metadata for the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: subordinateId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      requestBody:
        description: Metadata creation details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMetadata'
      responses:
        '201':
          description: Subordinate metadata created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateMetadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{subordinateId}/metadata/{id}':
    delete:
      summary: Delete subordinate metadata
      description: Delete metadata associated with the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: subordinateId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the metadata to be deleted.
      responses:
        '200':
          description: Subordinate metadata deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateMetadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subordinates/{id}/statement':
    get:
      summary: Get subordinate statement object
      description: Retrieve the subordinate statement for the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      responses:
        '200':
          description: Successfully retrieved subordinate statement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateStatement'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Sign and publish subordinate statement
      description: Sign and publish the subordinate statement for the specified subordinate. If `dry-run` is `true`, outputs the signed JWT without publishing it.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
      requestBody:
        description: Dry-run flag (optional)
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishEntityStatement'
      responses:
        '200':
          description: Subordinate statement published successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    AccountUsername:
      name: X-Account-Username
      in: header
      description: Optional account username to operate on behalf of
      required: false
      schema:
        type: string
...
  schemas:
...
    BaseJwk:
      type: object
      x-tags:
        - federation
      required:
        - kty
        - kid
      properties:
        kty:
          type: string
          description: 'The key type (e.g., EC).'
          example: EC
        crv:
          type: string
          description: The elliptic curve used (only for EC keys).
          example: P-256
          nullable: true
        kid:
          type: string
          description: The key ID.
          example: 12345
          nullable: true
        x:
          type: string
          description: The X coordinate for EC keys (optional).
          example: o-7zraXKDaoBte2PsuTXo-MSLzsyWdAElNptGgI4aH8
          nullable: true
        y:
          type: string
          description: The Y coordinate for EC keys (optional).
          example: Xr_wCzJ1XnsgAIV5qHruzSwaNnwy87UjmevVklTpIv8
          nullable: true
        n:
          type: string
          description: The modulus for RSA keys.
          example: modulus_value
          nullable: true
        e:
          type: string
          description: The exponent for RSA keys.
          example: AQAB
          nullable: true
        alg:
          type: string
          description: The algorithm associated with the key.
          example: ES256
          nullable: true
        use:
          type: string
          description: 'The intended use of the key (e.g., sig, enc).'
          example: sig
          nullable: true
        x5u:
          type: string
          format: uri
          description: A URL that points to an X.509 public key certificate or certificate chain.
          example: 'https://example.com/cert.pem'
          nullable: true
        x5c:
          type: array
          description: The X.509 certificate chain.
          nullable: true
          items:
            type: string
            description: A base64-encoded string representing an X.509 certificate.
            example: MIICoTCCAYkCAQ...
        x5t:
          type: string
          description: The SHA-1 thumbprint of the X.509 certificate.
          example: dGhpcyBpcyBqdXN0IGEgdGh1bWJwcmludA
          nullable: true
        x5tS256:
          type: string
          description: The SHA-256 thumbprint of the X.509 certificate.
          example: sM4KhEI1Y2Sb6-EVr6tJabmJuoP-ZE...
          nullable: true
    BaseStatement:
      type: object
      required:
        - iss
        - sub
        - iat
        - exp
        - jwks
      properties:
        iss:
          type: string
          description: The Entity Identifier of the issuer of the Entity Statement.
        sub:
          type: string
          description: The Entity Identifier of the subject.
        exp:
          type: integer
          description: Expiration time after which the statement MUST NOT be accepted for processing.
        iat:
          type: integer
          format: date-time
          description: The time the statement was issued.
        jwks:
          type: object
          x-tags:
            - federation
          properties:
            keys:
              type: array
              items:
                $ref: '#/components/schemas/Jwk'
        metadata:
          additionalProperties: true
        crit:
          type: array
          items:
            type: string
...
    CreateMetadata:
      type: object
      properties:
        key:
          type: string
          description: The metadata key.
          example: openid_relying_party
        metadata:
          additionalProperties: true
          description: The metadata object.
      required:
        - key
        - metadata
...
    CreateSubordinate:
      type: object
      properties:
        identifier:
          type: string
          description: The identifier of the subordinate account.
          example: 'https://www.example.com/subordinate'
      required:
        - identifier
...
    ErrorResponse:
      type: object
      x-tags:
        - federation
      required:
        - error
        - message
      properties:
        status:
          type: integer
          x-stoplight:
            id: joax1cgth4uzd
        error:
          type: string
        message:
          type: string
          description: A human-readable short text describing the error.
        timestamp:
          type: integer
          x-stoplight:
            id: qtn6mqbzpjctv
...
    Jwk:
      allOf:
        - $ref: '#/components/schemas/BaseJwk'
        - type: object
          x-tags:
            - federation
          properties:
            id:
              type: integer
              description: The unique identifier for the JWK record.
              example: 1
            uuid:
              type: string
              format: uuid
              description: The universally unique identifier for the JWK record.
              example: 123e4567-e89b-12d3-a456-426614174000
            account_id:
              type: integer
              description: The ID of the account associated with this JWK.
              example: 100
            revoked_at:
              type: string
              format: date-time
              description: 'The timestamp when the JWK was revoked, if applicable.'
              example: '2024-09-01T12:34:56Z'
              nullable: true
            revoked_reason:
              type: string
              description: 'The reason for revoking the JWK, if applicable.'
              example: Key compromise
              nullable: true
            created_at:
              type: string
              format: date-time
              description: The timestamp when the JWK was created.
              example: '2024-08-06T12:34:56Z'
              nullable: true
...
    Metadata:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the Metadata record.
          example: 1
        account_id:
          type: integer
          description: The ID of the account associated with this Metadata.
          example: 1
        key:
          type: string
          description: The key of the metadata.
          example: openid_relying_party
        metadata:
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was created.
          example: '2024-08-06T12:34:56Z'
          nullable: false
        deleted_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was deleted.
          example: '2024-08-06T12:34:56Z'
          nullable: true
    PublishEntityStatement:
      type: object
      x-tags:
        - federation
      properties:
        dry-run:
          type: boolean
          description: 'If true, the request will be validated but not persisted.'
          example: false
...
    Subordinate:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The unique identifier of the subordinate.
          example: 123
        accountId:
          type: integer
          format: int32
          description: The ID of the account associated with this subordinate.
          example: 456
        identifier:
          type: string
          description: The unique identifier for the subordinate.
          example: 'https://www.example.com/subordinate'
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the subordinate was created.
          example: '2023-08-21T14:52:00Z'
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: 'The timestamp when the subordinate was deleted, if applicable.'
          example: '2024-08-21T14:52:00Z'
      required:
        - id
        - accountId
        - identifier
        - createdAt
    SubordinateJwk:
      type: object
      x-tags:
        - federation
      properties:
        id:
          type: integer
          description: The unique identifier for the Subordinate key record.
          example: 1
        subordinate_id:
          type: integer
          description: The ID of the subordinated account associated with this key.
          example: 1
        key:
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: The timestamp when the key  was created.
          example: '2024-08-06T12:34:56Z'
          nullable: false
    SubordinateMetadata:
      type: object
      x-tags:
        - federation
      properties:
        id:
          type: integer
          description: The unique identifier for the Subordinate Metadata record.
          example: 1
        account_id:
          type: integer
          description: The ID of the account associated with this Metadata.
          example: 1
        subordinate_id:
          type: integer
          description: The ID of the subordinate associated with this Metadata.
          example: 1
        key:
          type: string
          description: The key of the metadata.
          example: openid_relying_party
        metadata:
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was created.
          example: '2024-08-06T12:34:56Z'
          nullable: false
        deleted_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was deleted.
          example: '2024-08-06T12:34:56Z'
          nullable: true
    SubordinateStatement:
      allOf:
        - $ref: '#/components/schemas/BaseStatement'
        - type: object
          required:
            - iss
            - sub
            - iat
            - exp
            - jwks
          properties:
            metadata_policy:
              additionalProperties: true
            crit:
              type: array
              items:
                type: string
            metadata_policy_crit:
              additionalProperties: true
            source_endpoint:
              type: string
              format: uri
              description: String containing the fetch endpoint URL from which the Entity Subordinate Statement was issued.
...